[Feature] CELONIS_APPLY_ABC_MODEL (change to return BIGINT) (#782)
Change the return type from INT to BIGINT.

Tested locally.

- [ ] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
- [ ] I have added documentation for my new feature or new function

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4

+++ gensrc/script/functions.py
@@ -1514,8 +1514,8 @@ celostar_functions = [
     [5000103, 'celonis_square', True, False, 'DOUBLE', ['DOUBLE'], 'CelonisMathFunctions<TYPE_DOUBLE>::square'],
     [5000200, 'celonis_qnorm', True, False, 'DOUBLE', ['DOUBLE'], 'CelonisQnorm::qnorm'],
     [5000300, 'celonis_predict_linear_regression', True, False, 'DOUBLE', ['ARRAY_DOUBLE', 'VARCHAR'], 'CelonisLinearRegression::predict_linear_regression', 'CelonisLinearRegression::predict_prepare', 'CelonisLinearRegression::predict_close'],
-    [5000401, 'celonis_apply_abc_model', True, False, 'INT', ['BIGINT', 'BIGINT', 'VARCHAR'], 'CelonisAbcModel<TYPE_BIGINT>::apply_abc_model', 'CelonisAbcModel<TYPE_BIGINT>::prepare', 'CelonisAbcModel<TYPE_BIGINT>::close'],
-    [5000402, 'celonis_apply_abc_model', True, False, 'INT', ['DOUBLE', 'BIGINT', 'VARCHAR'], 'CelonisAbcModel<TYPE_DOUBLE>::apply_abc_model', 'CelonisAbcModel<TYPE_DOUBLE>::prepare', 'CelonisAbcModel<TYPE_DOUBLE>::close'],
+    [5000401, 'celonis_apply_abc_model', True, False, 'BIGINT', ['BIGINT', 'BIGINT', 'VARCHAR'], 'CelonisAbcModel<TYPE_BIGINT>::apply_abc_model', 'CelonisAbcModel<TYPE_BIGINT>::prepare', 'CelonisAbcModel<TYPE_BIGINT>::close'],
+    [5000402, 'celonis_apply_abc_model', True, False, 'BIGINT', ['DOUBLE', 'BIGINT', 'VARCHAR'], 'CelonisAbcModel<TYPE_DOUBLE>::apply_abc_model', 'CelonisAbcModel<TYPE_DOUBLE>::prepare', 'CelonisAbcModel<TYPE_DOUBLE>::close'],
 ]
 
 vectorized_functions.extend(celostar_functions)
